dblib: Don't raise exceptions from C callbacks
Raising an exception from the message or error handler destroys the C
call stack, which does unexpected things to FreeTDS. For example, if
an error occurs and you throw an exception in the error handler,
FreeTDS will continue re-notifying that error on every subsequent
call.

To fix this, catch any exceptions in the message handler, store them
in the FreeTDS userdata, then rethrow the exception once the FreeTDS
call finishes.

(Note that we still throw exceptions immediately if a function
isn't associated with a DBPROCESS).

This requires tedious calls to check if we stored an exception after
every FreeTDS call, but I can't think of a better way to do this.